requirements.txt
pandas
requests
numpy
psycopg2-binary
fastapi
uvicorn
openpyxl
